fix(StopDetailsViewModel): Pull departures in to prevent flickering StopDetailsPage
