+++ androidApp/src/main/java/com/mbta/tid/mbta_app/android/pages/StopDetailsPage.kt
@@ -8,19 +8,12 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import com.mapbox.maps.MapboxExperimental
 import com.mbta.tid.mbta_app.android.component.ErrorBannerViewModel
-import com.mbta.tid.mbta_app.android.state.getGlobalData
 import com.mbta.tid.mbta_app.android.stopDetails.StopDetailsView
 import com.mbta.tid.mbta_app.android.stopDetails.StopDetailsViewModel
 import com.mbta.tid.mbta_app.android.util.managePinnedRoutes
-import com.mbta.tid.mbta_app.android.util.rememberSuspend
-import com.mbta.tid.mbta_app.android.util.timer
 import com.mbta.tid.mbta_app.model.StopDetailsDepartures
 import com.mbta.tid.mbta_app.model.StopDetailsFilter
 import com.mbta.tid.mbta_app.model.StopDetailsPageFilters
-import com.mbta.tid.mbta_app.model.response.AlertsStreamDataResponse
-import kotlin.time.Duration.Companion.seconds
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
 
 @Composable
 @ExperimentalMaterial3Api
@@ -29,51 +22,16 @@ fun StopDetailsPage(
     modifier: Modifier = Modifier,
     viewModel: StopDetailsViewModel,
     filters: StopDetailsPageFilters,
-    alertData: AlertsStreamDataResponse?,
     onClose: () -> Unit,
     updateStopFilter: (StopDetailsFilter?) -> Unit,
     updateDepartures: (StopDetailsDepartures?) -> Unit,
     errorBannerViewModel: ErrorBannerViewModel
 ) {
-    val globalResponse = getGlobalData("StopDetailsPage.getGlobalData")
-
     val stopId = filters.stopId
-
-    val now = timer(updateInterval = 5.seconds)
-
-    val stopData by viewModel.stopData.collectAsState(null)
-
-    val schedulesResponse = stopData?.schedules
-
-    val predictionsResponse = stopData?.predictionsByStop?.toPredictionsStreamDataResponse()
 
     val (pinnedRoutes, togglePinnedRoute) = managePinnedRoutes()
 
-    val departures =
-        rememberSuspend(
-            stopId,
-            globalResponse,
-            schedulesResponse,
-            stopId,
-            alertData,
-            pinnedRoutes,
-            now
-        ) {
-            withContext(Dispatchers.Default) {
-                if (globalResponse != null) {
-                    StopDetailsDepartures.fromData(
-                        stopId,
-                        globalResponse,
-                        schedulesResponse,
-                        predictionsResponse,
-                        alertData,
-                        pinnedRoutes.orEmpty(),
-                        now,
-                        useTripHeadsigns = false,
-                    )
-                } else null
-            }
-        }
+    val departures by viewModel.stopDepartures.collectAsState()
 
     LaunchedEffect(departures) { updateDepartures(departures) }
